third_party/WebKit/LayoutTests/editing/spelling/spellcheck-editable-on-focus-multiframe-expected.txt

Spell checking should be triggered on focus of an editable. Also after retarting spell checker. It applies to all frames.To test manually type focus above editable. Misspellings in them should be marked on focus.Turn spell checker off and on again and again focus editables. Mispelling should be marked again

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.markerCountForNode(testEditable1.childNodes[0], "spelling") is 0
PASS internals.markerCountForNode(testEditable2.childNodes[0], "spelling") is 0
PASS internals.markerCountForNode(testEditable1.childNodes[0], "spelling") is 3
PASS internals.markerCountForNode(testEditable1.childNodes[0], "spelling") is 3
PASS internals.markerCountForNode(testEditable1.childNodes[0], "spelling") is 0
PASS internals.markerCountForNode(testEditable2.childNodes[0], "spelling") is 0
PASS internals.markerCountForNode(testEditable1.childNodes[0], "spelling") is 3
PASS internals.markerCountForNode(testEditable2.childNodes[0], "spelling") is 3
PASS successfullyParsed is true

TEST COMPLETE
 
